tcgbios: Require user interaction for executing TPM PPI operations
Require user confirmation for the execution of a TPM PPI operation.

The processing of the TPM PPI operations can happen without user
interaction but requires the implementation of another PPI command to
request the execution of command without interaction. However, this in
turn requires some persistent storage where the configuration can be
stored. Such persistent storage could be defined in a TPM NVRAM space
in a future extension.

Since confirmation requires user interaction we have to move the call to
tpm_ppi_process() to a point in time after the screen and keyboard have
been initialized.

The specs suggest the usage of F10 and F12 keys for acceptance to execute
the TPM operation. However, for virtual machines accessed over a VNC
session this becomes more difficult since those keys may be intercepted
by the host OS and not reach the VM. Therefore, we use the 'A' and 'C'
keys instead.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>